backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
email-validator>=2.0
python-multipart>=0.0.9
passlib>=1.7.4
python-jose[cryptography]>=3.3.0

